+++ tests/Bridge/Elasticsearch/DataProvider/PaginatorTest.php
@@ -166,7 +166,7 @@ class PaginatorTest extends TestCase
         );
     }
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->paginator = $this->getPaginator($this->limit, $this->offset);
     }