+++ tests/Bridge/Symfony/Bundle/Test/ApiTestCaseTest.php
@@ -1,0 +1,85 @@
+<?php
+
+/*
+ * This file is part of the API Platform project.
+ *
+ * (c) Kévin Dunglas <dunglas@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace ApiPlatform\Core\Tests\Bridge\Symfony\Bundle\Test;
+
+use ApiPlatform\Core\Bridge\Symfony\Bundle\Test\ApiTestCase;
+use PHPUnit\Framework\ExpectationFailedException;
+use PHPUnit\Runner\Version;
+
+class ApiTestCaseTest extends ApiTestCase
+{
+    public function testAssertJsonContains(): void
+    {
+        if (version_compare(Version::id(), '8.0.0', '<')) {
+            $this->markTestSkipped('Requires PHPUnit 8');
+        }
+
+        self::createClient()->request('GET', '/');
+        $this->assertJsonContains(['@context' => '/contexts/Entrypoint']);
+    }
+
+    public function testAssertJsonEquals(): void
+    {
+        self::createClient()->request('GET', '/contexts/Address');
+        $this->assertJsonEquals([
+            '@context' => [
+                '@vocab' => 'http://example.com/docs.jsonld#',
+                'hydra' => 'http://www.w3.org/ns/hydra/core#',
+                'name' => 'Address/name',
+            ],
+        ]);
+    }
+
+    public function testAssertMatchesJsonSchema(): void
+    {
+        $jsonSchema = <<<JSON
+{
+  "type": "object",
+  "properties": {
+    "@context": {"pattern": "^/contexts/Entrypoint"},
+    "@id": {"pattern": "^/$"},
+    "@type": {"pattern": "^Entrypoint$"},
+    "dummy": {}
+  },
+  "additionalProperties": true,
+  "required": ["@context", "@id", "@type", "dummy"]
+}
+JSON;
+
+        self::createClient()->request('GET', '/');
+        $this->assertMatchesJsonSchema($jsonSchema);
+        $this->assertMatchesJsonSchema(json_decode($jsonSchema, true));
+    }
+
+    // Next tests have been imported from dms/phpunit-arraysubset-asserts, because the original constraint has been deprecated.
+
+    public function testAssertArraySubsetPassesStrictConfig(): void
+    {
+        if (version_compare(Version::id(), '8.0.0', '<')) {
+            $this->markTestSkipped('Requires PHPUnit 8');
+        }
+
+        $this->expectException(ExpectationFailedException::class);
+        $this->assertArraySubset(['bar' => 0], ['bar' => '0'], true);
+    }
+
+    public function testAssertArraySubsetDoesNothingForValidScenario(): void
+    {
+        if (version_compare(Version::id(), '8.0.0', '<')) {
+            $this->markTestSkipped('Requires PHPUnit 8');
+        }
+
+        $this->assertArraySubset([1, 2], [1, 2, 3]);
+    }
+}