+++ tests/Fixtures/TestBundle/Document/DummyCustomQuery.php
@@ -29,8 +29,27 @@ use Doctrine\ODM\MongoDB\Mapping\Annotations as ODM;
  *         "item_query"="app.graphql.query_resolver.dummy_custom_not_retrieved_item_document",
  *         "args"={}
  *     },
+ *     "testItemCustomArguments"={
+ *         "item_query"="app.graphql.query_resolver.dummy_custom_item",
+ *         "args"={
+ *             "id"={"type"="ID"},
+ *             "customArgumentNullableBool"={"type"="Boolean"},
+ *             "customArgumentBool"={"type"="Boolean!"},
+ *             "customArgumentInt"={"type"="Int!"},
+ *             "customArgumentString"={"type"="String!"},
+ *             "customArgumentFloat"={"type"="Float!"},
+ *             "customArgumentIntArray"={"type"="[Int!]!"},
+ *             "customArgumentCustomType"={"type"="DateTime!"}
+ *         }
+ *     },
  *     "testCollection"={
  *         "collection_query"="app.graphql.query_resolver.dummy_custom_collection"
+ *     },
+ *     "testCollectionCustomArguments"={
+ *         "collection_query"="app.graphql.query_resolver.dummy_custom_collection",
+ *         "args"={
+ *             "customArgumentString"={"type"="String!"}
+ *         }
  *     }
  * })
  * @ODM\Document
@@ -48,4 +67,9 @@ class DummyCustomQuery
      * @var string
      */
     public $message;
+
+    /**
+     * @var array
+     */
+    public $customArgs = [];
 }